fix : rename styling files as per convention
