+++ src/Components/CoursePage/CourseContent/MyAccordion/MyAccordion.js
@@ -3,7 +3,7 @@ import Accordion from "@mui/material/Accordion";
 import AccordionSummary from "@mui/material/AccordionSummary";
 import AccordionDetails from "@mui/material/AccordionDetails";
 import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
-import myAcordionStyles from "./myAcordionStyles.module.css";
+import myAcordionStyles from "./MyAccordion.module.css";
 
 import AccordionItem from "./AccordionItem/AccordionItem";
 function MyAccordion(props) {
