+++ src/Components/CoursePage/FixedCard/FixedCard.module.css
