+++ src/Components/CoursePage/ShortcutNav/ShortcutNav.module.css
