+++ src/Components/CoursePage/StudentFeedback/StudentFeedback.js
@@ -1,9 +1,9 @@
 import React, { useContext } from "react";
 import { BsStarFill, BsStarHalf, BsStar } from "react-icons/bs";
 import { useParams } from "react-router-dom";
-import studentFeedbackStyles from "./studentFeedbackStyles.module.css";
+import studentFeedbackStyles from "./StudentFeedback.module.css";
 import { ReviewsContext } from "../../../App";
-import courseContentStyles from "/Users/abdullahahmad/my-app/src/Components/CoursePage/CourseContent/courseContentStyles.module.css";
+import courseContentStyles from "/Users/abdullahahmad/my-app/src/Components/CoursePage/CourseContent/CourseContent.module.css";
 
 function StudentFeedback() {
   const params = useParams();
